requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
numpy==1.26.4
pandas==2.2.2
Pillow>=11.0.0
opencv-python==4.9.0.80
scikit-learn==1.4.2
joblib==1.4.2
matplotlib==3.9.0
xgboost==2.0.3
scipy>=1.11.0
openpyxl>=3.1.0
tensorflow-cpu>=2.11.0
